+++ components/frontend/src/fields/StringInput.test.js
@@ -1,46 +1,90 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { mount, shallow } from 'enzyme';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import { Permissions } from '../context/Permissions';
 import { StringInput } from './StringInput';
 
-it('renders without crashing', () => {
-    const div = document.createElement('div');
-    ReactDOM.render(<StringInput />, div);
-    ReactDOM.unmountComponentAtNode(div);
+function renderStringInput(set_value) {
+    return render(
+        <Permissions.Provider value={false}>
+            <StringInput
+                options={["Option 1", "Option 2"]}
+                set_value={set_value}
+                value="Option 1"
+            />
+        </Permissions.Provider>
+    )
+}
+
+it('renders the value of the input', () => {
+    renderStringInput();
+    expect(screen.getByDisplayValue(/Option 1/)).not.toBe(null)
 });
 
-it('renders the value of the Input', () => {
-    const wrapper = shallow(<StringInput value="Hello" />);
-    expect(wrapper.find('Input').prop('value')).toBe("Hello");
+it('renders a missing value', () => {
+    render(<StringInput options={["Option 1", "Option 2"]} />)
+    expect(screen.queryByDisplayValue(/Option/)).toBe(null)
 });
 
-it('renders the value of the StringInputWithSuggestions', () => {
-    const wrapper = mount(<Permissions.Provider value={false}><StringInput value="Hello" options='x' /></Permissions.Provider>);
-    expect(wrapper.find('StringInputWithSuggestions').prop('value')).toBe("Hello");
+it('invokes the callback on change', () => {
+    const mockCallback = jest.fn();
+    renderStringInput(mockCallback);
+    userEvent.type(screen.getByDisplayValue(/Option 1/), '{selectall}Option 2{enter}')
+    expect(screen.getByDisplayValue(/Option 2/)).not.toBe(null)
+    expect(mockCallback).toHaveBeenCalledWith("Option 2")
 });
 
-it('renders error prop to true when warning is set to true', () => {
-    const wrapper = mount(<Permissions.Provider value={false}><StringInput warning={true} options='x' /></Permissions.Provider>);
-    expect(wrapper.find('FormDropdown').prop('error')).toBe(true);
+it('invokes the callback on add', () => {
+    const mockCallback = jest.fn();
+    renderStringInput(mockCallback);
+    userEvent.type(screen.getByDisplayValue(/Option 1/), '{selectall}Option 3{enter}')
+    expect(screen.getByDisplayValue(/Option 3/)).not.toBe(null)
+    expect(mockCallback).toHaveBeenCalledWith("Option 3")
 });
 
-it('renders error prop to true when the field is required and empty and the warning is undefined', () => {
-    const wrapper = mount(<Permissions.Provider value={false}><StringInput required={true} value='' options='x' /></Permissions.Provider>);
-    expect(wrapper.find('FormDropdown').prop('error')).toBe(true);
+it('does not invoke the callback when the new value equals the old value', () => {
+    const mockCallback = jest.fn();
+    renderStringInput(mockCallback);
+    userEvent.type(screen.getByDisplayValue(/Option 1/), '{selectall}Option 1{enter}')
+    expect(screen.getByDisplayValue(/Option 1/)).not.toBe(null)
+    expect(mockCallback).not.toHaveBeenCalled()
 });
 
-it('renders error prop to false when the field is required and not empty and the warning is undefined', () => {
-    const wrapper = mount(<Permissions.Provider value={false}><StringInput required={true} value='x' options='x' /></Permissions.Provider>);
-    expect(wrapper.find('FormDropdown').prop('error')).toBe(false);
+it('works without options', () => {
+    const mockCallback = jest.fn();
+    renderStringInput(mockCallback);
+    render(<StringInput set_value={mockCallback} />)
+    userEvent.type(screen.getByDisplayValue(""), '{selectall}New value{enter}')
+    expect(screen.getByDisplayValue(/New value/)).not.toBe(null)
+    expect(mockCallback).toHaveBeenCalledWith("New value")
 });
 
-it('renders error prop to true when the field is required and empty and the warning is false', () => {
-    const wrapper = mount(<Permissions.Provider value={false}><StringInput required={true} warning={false} value='' options='x' /></Permissions.Provider>);
-    expect(wrapper.find('FormDropdown').prop('error')).toBe(true);
+it('shows an error for required empty fields', () => {
+    render(<StringInput options={["Option 1", "Option 2"]} required />)
+    expect(screen.getByRole("combobox")).toBeInvalid()
 });
 
-it('renders error prop to false when the field is required and not empty and the warning is false', () => {
-    const wrapper = mount(<Permissions.Provider value={false}><StringInput required={true} warning={false} value='x' options='x' /></Permissions.Provider>);
-    expect(wrapper.find('FormDropdown').prop('error')).toBe(false);
+it('does not show an error for required non-empty fields', () => {
+    render(<StringInput options={["Option 1", "Option 2"]} value="Hello" required />)
+    expect(screen.getByRole("combobox")).toBeValid()
+});
+
+it('does not show an error for non-required empty fields', () => {
+    render(<StringInput options={["Option 1", "Option 2"]} value="" />)
+    expect(screen.getByRole("combobox")).toBeValid()
+});
+
+it('does not show an error for non-required non-empty fields', () => {
+    render(<StringInput options={["Option 1", "Option 2"]} value="Hello" />)
+    expect(screen.getByRole("combobox")).toBeValid()
+});
+
+it('shows an error', () => {
+    render(<StringInput options={["Option 1", "Option 2"]} value="Hello" error />)
+    expect(screen.getByRole("combobox")).toBeInvalid()
+});
+
+it('shows a warning', () => {
+    render(<StringInput options={["Option 1", "Option 2"]} value="Hello" warning />)
+    expect(screen.getByRole("combobox")).toBeInvalid()
 });