+++ components/frontend/src/fields/TextInput.js
@@ -1,47 +1,64 @@
 import React, { useState } from 'react';
 import { Form } from 'semantic-ui-react';
-import { accessGranted, Permissions } from '../context/Permissions';
+import { ReadOnlyOrEditable } from '../context/Permissions';
 
-export function TextInput(props) {
-    let { editableLabel, required, set_value, requiredPermissions, ...otherProps } = props;
-    const [text, setText] = useState(props.value || '');
-    function readOnly(permissions) {
-        return requiredPermissions ? !accessGranted(permissions, requiredPermissions) : false
-    }
-    function onChange(event) {
-        setText(event.target.value)
-    }
+function ReadOnlyTextInput({ label, required, value }) {
+    return (
+        <Form>
+            <Form.TextArea
+                error={required && value === ""}
+                label={label}
+                readOnly
+                tabIndex={-1}
+                value={value}
+            />
+        </Form>
+    )
+}
+
+function EditableTextInput(props) {
+    let { editableLabel, label, required, set_value, ...otherProps } = props;
+    const initialValue = props.value || ""
+    const [text, setText] = useState(initialValue);
+
     function onKeyDown(event) {
         if (event.key === "Escape") {
-            setText(props.value || '')
+            setText(initialValue)
         }
     }
     function onKeyPress(event) {
-        if (event.key === "Enter" && event.shiftKey && text !== (props.value || '')) {
+        if (event.key === "Enter" && event.shiftKey) {
             event.preventDefault();
-            props.set_value(text);
+            submit()
         }
     }
     function submit() {
-        if (text !== (props.value || '')) {
-            props.set_value(text)
+        if (text !== initialValue) {
+            set_value(text)
         }
     }
     return (
         <Form onSubmit={submit}>
-            <Permissions.Consumer>{(permissions) => (
-                <Form.TextArea
-                    {...otherProps}
-                    error={required && text === ""}
-                    label={readOnly(permissions) ? editableLabel || props.label : props.label}
-                    onBlur={submit}
-                    onChange={onChange}
-                    onKeyDown={onKeyDown}
-                    onKeyPress={onKeyPress}
-                    readOnly={readOnly(permissions)}
-                    value={text}
-                />)}
-            </Permissions.Consumer>
+            <Form.TextArea
+                {...otherProps}
+                error={required && text === ""}
+                label={editableLabel || label}
+                onBlur={submit}
+                onChange={(event) => setText(event.target.value)}
+                onKeyDown={onKeyDown}
+                onKeyPress={onKeyPress}
+                value={text}
+            />
         </Form>
+    )
+}
+
+export function TextInput(props) {
+    let { requiredPermissions, ...otherProps } = props;
+    return (
+        <ReadOnlyOrEditable
+            requiredPermissions={requiredPermissions}
+            readOnlyComponent={<ReadOnlyTextInput {...otherProps} />}
+            editableComponent={<EditableTextInput {...otherProps} />} />
     )
 }